Removed testing classes
